Execute callback in background.js for a Firefox module event
[Reference](https://firefox-source-docs.mozilla.org/toolkit/components/extensions/webextensions/reference.html#eventmanager-class)
Thanks to aswan and Gijs for their assistance.

This WebExtension Experiment executes a callback in the background script when a Firefox module event fires. In this example the Firefox module is CustomizeUI.jsm, and the event is onWidgetRemoved.

To test that this works, remove the browserAction button from the toolbar in Firefox and observe the message logged in the Browser Console from 'background.js':
1. cd wee-detect-hide-toolbar-button
2. web-ext run
3. open the Browser Console in Firefox and filter for output from background.js
4. In the Firefox window, right click the browserAction icon for the extension in the toolbar area
5. Click 'Remove from Toolbar'
6. Notice a message is logged to the console: 'Widget removed: ${widgetId}', where widgetId is the id of the browserAction toolbar button removed.