Add browserAction popup
This is to see what happens when there is no background page (i.e. by removing the 'background' key from 'manifest.json') for an extension (say it just has a popup) when I open the Storage inspector.

Right now we are using the 'browser.storage' WE API directly off of the extension's 'window' object (e.g.:

```
  // TODO: Don't use WE APIs directly, as a WE can overwrite the `browser` object
  async removeItem(host, name) {
    const win = this.storageActor.getWindowFromHost(host);
    await Cu.waiveXrays(win).browser.storage.local.remove(name);
  },
```

...but `window` not always available as in the case when there is no background page running in the extension and no other extension pages are open. I have confirmed this does indeed break with the following error:

```
Protocol error (unknownError): Error occurred while creating actor' server2.conn0.child13/storageActor8: TypeError: Cu.waiveXrays(...).browser is undefined
```

...in ui.js:165:15. So we need to rely on something other than 'window' to get at the extension's storage.

It may be possible if we can get the addonId some way to access the extension's storage via the privileged implementation of the storage WE API.