Add extension page to use Browser Content Toolbox debugger
Because the storage actor (in mozilla-central/devtools/server/actors/storage.js) for Bug 1292234 is running in the extension child process, it cannot be debugged from the Browser Toolbox (debugger for main process), the addon debugger (debugger for web extension itself) or the in-content toolbox (debugger for web page scripts).

With Fission, there will be a single toolbox that can target any different process, but in the meantime, there is a workaround to be able to debug this storage actor with the Browser Content Toolbox.

1) Open a webextension page of the target extension in a Firefox tab
2) Open the “Browser Content Toolbox” (from App Menu -> “Web Developers”) while the webextension page tab is currently selected (this way a toolbox window connected to the extension child process should be opened)